requirements.txt
ai-microcore==4.3.4
fastmcp~=2.12.2
rich~=14.1.0